Add KafkaSink upgrade tests (#1716) (#70)
* Add KafkaSink upgrade tests

To implement KafkaSink upgrade tests (pre, post) we reuse an existing
utility that runs a KafkaSink test provided a KafkaSinkSpec object.

The continual test for KafkaSink is implemented using a combination
`KafkaSink` and `KafkaSource` as follows:

```
-> Upgrade test source
  -> `KafkaSink`
    -> `KafkaSource`
      -> Upgrade test sink
```

We use KafkaSink and KafkaSource to avoid creating a specific
kafka consumer for the upgrade test.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>

* Improve log format

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>