Reduce global resyncs for Source v2
In the soak tests, I observed that the queue depth was constantly very
high for the ConsumerGroup reconciler, which leads to a very slow
reconciliation process.

One reason for this might be that the scheduler is single thread but
the main reason is that we were doing a lot of global resyncs
(in both Consumer and ConsumerGroup reconciler) triggered by any
StatefulSet updates which produces many API server remote call and
have increased the API server remote calls latency due to client-side
throttling.

To avoid global resyncs we're now requeue-ing individual objects
only when some (expected) replicas are not schedule-able due to
missing pods, and we drop the need to global resync when the
statefulset changes in response to the autoscaler scaling up
the statefulset to make room for new replicas.

In the eviction case, when the statefulset is scaled down,
individual objects are patched, which triggers ConsumerGroup
reconciliation, so we need to queue individual objects only
when virtual replicas count is not equal to the expected
number of replicas.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>