Fix slow consumergroup reconciliation under load (#788)
In [1], we return an error, however, that method is called in
the consumer group reconciler when a consumer is not ready
which is a normal state at the beginning when consumers have been
just created, so we shouldn't return an error because that causes
the consumer group to be reconciled again with an exponentially
increasing delay causing slow time to ready.

This is especially evident when scaling up with a high load =
(therefore when dispatcher pod is slow to become ready).

[1] https://github.com/knative-sandbox/eventing-kafka-broker/blob/5cda5463aa2fa060179674fe7b3237abb836ee06/control-plane/pkg/apis/internals/kafka/eventing/v1alpha1/consumer_group_lifecycle.go#L57-L65

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>
Co-authored-by: Pierangelo Di Pilato <pierdipi@redhat.com>