Remove unschedulable pods marking during eviction
This is racing with the autoscaler, so we cannot really
make pods as unschedulable.

I think this was an early optimization to avoid many
rescheduling because the risk is that we end up rescheduling
many consumer, however, to mitigate this we can also think about
increasing the "scale down refresh period" so that we don't
scale up / down for at least that time window.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>